moved databaseIndex and query code to ChildrenService
so that it can be reused by other components easily
