+++ src/app/children/attendance/attendance-average-dashboard/attendance-average-dashboard.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import {ChildrenService} from '../../children.service';
-import {Database} from '../../../database/database';
 import {Child} from '../../child';
 import {Router} from '@angular/router';
 import {AttendanceMonth} from '../attendance-month';
@@ -16,84 +15,18 @@ export class AttendanceAverageDashboardComponent implements OnInit {
   overallAttendance: number;
   lastMonthsTopAttendence = []; // [[Child, average_last_3_months, last_months_attendance]]
 
-  constructor(private db: Database,
-              private childrenService: ChildrenService,
+  constructor(private childrenService: ChildrenService,
               private router: Router) { }
 
   ngOnInit() {
-    this.createAttendanceAnalysisIndex()
-      .then(() => this.loadAverageAttendances());
+    this.loadAverageAttendances();
   }
 
-  private createAttendanceAnalysisIndex(): Promise<any> {
-    const emit = (x, y?) => {}; // defined to avoid Typescript error. Actually `emit` is provided by pouchDB to the `map` function
-
-    const designDoc = {
-      _id: '_design/avg_attendance_index',
-      views: {
-        three_months: {
-          map: averageAttendanceMap.toString(),
-          reduce: '_stats'
-        },
-        last_month: {
-          map: lastAverageAttendanceMap.toString(),
-          reduce: '_stats'
-        }
-      }
-    };
-
-    return this.db.saveDatabaseIndex(designDoc);
-
-
-    // `emit(x)` to add x as a key to the index that can be searched
-    function averageAttendanceMap (doc) {
-      if (!doc._id.startsWith('AttendanceMonth:')) {
-        return;
-      }
-      if (!isWithinLast3Months(new Date(doc.month), new Date())) {
-        return;
-      }
-
-      emit(doc.student, doc.daysAttended / (doc.daysWorking - doc.daysExcused));
-
-      function isWithinLast3Months(date: Date, now: Date) {
-        let months;
-        months = (now.getFullYear() - date.getFullYear()) * 12;
-        months -= date.getMonth();
-        months += now.getMonth();
-
-        if (months < 0) {
-          return false;
-        }
-        return months <= 3;
-      }
-    }
-
-    function lastAverageAttendanceMap (doc) {
-      if (!doc._id.startsWith('AttendanceMonth:')) {
-        return;
-      }
-      if (!isWithinLastMonth(new Date(doc.month), new Date())) {
-        return;
-      }
-
-      emit(doc.student, doc.daysAttended / (doc.daysWorking - doc.daysExcused));
-
-      function isWithinLastMonth(date: Date, now: Date) {
-        let months;
-        months = (now.getFullYear() - date.getFullYear()) * 12;
-        months -= date.getMonth();
-        months += now.getMonth();
-
-        return months === 1;
-      }
-    }
-  }
 
 
   loadAverageAttendances() {
     const countMap = new Map<string, [Child, number, number]>();
-    this.db.query('avg_attendance_index/three_months', {reduce: true, group: true})
+    this.childrenService.queryAttendanceLast3Months()
       .then(queryResults => {
         let totalCount = 0;
         let summedAverage = 0;
@@ -106,7 +39,7 @@ export class AttendanceAverageDashboardComponent implements OnInit {
 
         this.overallAttendance = summedAverage / totalCount;
       })
-      .then(() => this.db.query('avg_attendance_index/last_month', {reduce: true, group: true}))
+      .then(() => this.childrenService.queryAttendanceLastMonth())
       .then(queryResults => {
         queryResults.rows.forEach(studentStat => {
           const record = countMap.get(studentStat.key);
