+++ src/app/children/children.service.ts
@@ -3,10 +3,15 @@ import {Observable} from 'rxjs/Rx';
 import {Child} from './child';
 import {EntityMapperService} from '../entity/entity-mapper.service';
 import {AttendanceMonth} from './attendance/attendance-month';
+import {Database} from '../database/database';
 
 @Injectable()
 export class ChildrenService {
-  constructor(private entityMapper: EntityMapperService) {
+  attendanceIndicesUpdated: Promise<any>;
+
+  constructor(private entityMapper: EntityMapperService,
+              private db: Database) {
+    this.attendanceIndicesUpdated = this.createAttendanceAnalysisIndex();
   }
 
   getChildren(): Observable<Child[]> {
@@ -41,4 +46,88 @@ export class ChildrenService {
   removeAttendance(entity: AttendanceMonth) {
     this.entityMapper.remove(entity);
   }
+
+
+  queryAttendanceLast3Months() {
+    return this.attendanceIndicesUpdated
+      .then(() => this.db.query('avg_attendance_index/three_months', {reduce: true, group: true}));
+  }
+
+  queryAttendanceLastMonth() {
+    return this.attendanceIndicesUpdated
+      .then(() => this.db.query('avg_attendance_index/last_month', {reduce: true, group: true}))
+  }
+
+
+  private createAttendanceAnalysisIndex(): Promise<any> {
+    const designDoc = {
+      _id: '_design/avg_attendance_index',
+      views: {
+        three_months: {
+          map: this.getAverageAttendanceMap().toString(),
+          reduce: '_stats'
+        },
+        last_month: {
+          map: this.getLastAverageAttendanceMap().toString(),
+          reduce: '_stats'
+        }
+      }
+    };
+
+    return this.db.saveDatabaseIndex(designDoc);
+  }
+
+  private getAverageAttendanceMap (): (doc) => void {
+    const emit = (x, y?) => {}; // defined to avoid Typescript error. Actually `emit` is provided by pouchDB to the `map` function
+    // `emit(x)` to add x as a key to the index that can be searched
+
+    return (doc) => {
+      if (!doc._id.startsWith('AttendanceMonth:')) {
+        return;
+      }
+      if (!isWithinLast3Months(new Date(doc.month), new Date())) {
+        return;
+      }
+
+      emit(doc.student, doc.daysAttended / (doc.daysWorking - doc.daysExcused));
+
+      function isWithinLast3Months(date: Date, now: Date) {
+        let months;
+        months = (now.getFullYear() - date.getFullYear()) * 12;
+        months -= date.getMonth();
+        months += now.getMonth();
+
+        if (months < 0) {
+          return false;
+        }
+        return months <= 3;
+      }
+    };
+  }
+
+  private getLastAverageAttendanceMap () {
+    const emit = (x, y?) => {}; // defined to avoid Typescript error. Actually `emit` is provided by pouchDB to the `map` function
+    // `emit(x)` to add x as a key to the index that can be searched
+
+    return (doc) => {
+      if (!doc._id.startsWith('AttendanceMonth:')) {
+        return;
+      }
+      if (!isWithinLastMonth(new Date(doc.month), new Date())) {
+        return;
+      }
+
+      emit(doc.student, doc.daysAttended / (doc.daysWorking - doc.daysExcused));
+
+      function isWithinLastMonth(date: Date, now: Date) {
+        let months;
+        months = (now.getFullYear() - date.getFullYear()) * 12;
+        months -= date.getMonth();
+        months += now.getMonth();
+
+        return months === 1;
+      }
+    };
+  }
+
 }
